pyproject.toml
[project]
name = "whisper-typing"
version = "1.0.0"
description = "A powerful, background speech-to-text application for Windows that runs locally"
readme = "README.md"
requires-python = ">=3.13.7"
license = "MIT"
authors = [
  { name = "Roger P. Filomeno", email = "rpfilomeno@gmail.com" },
]
keywords = ["example", "fibonacci"]
classifiers = [
  "Development Status :: 1 - Planning",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "accelerate>=1.12.0",
  "colorama>=0.4.6",
  "faster-whisper>=1.0.3",
  "google-api-core>=2.29.0",
  "google-genai>=1.59.0",
  "huggingface_hub[hf_xet]>=0.20.0",
  "numpy>=2.4.1",
  "optimum>=2.1.0",
  "pygetwindow>=0.0.9",
  "pynput>=1.8.1",
  "pyperclip>=1.11.0",
  "python-dotenv>=1.0.1",
  "scipy>=1.17.0",
  "sounddevice>=0.5.3",
  "textual>=0.70.0",
  "torchaudio>=2.0.0",
  "transformers>=4.57.6",
]

[project.urls]
issues = "https://github.com/rpfilomeno/whispher-typing/issues"
repository = "https://github.com/rpfilomeno/whispher-typing"

[project.scripts]
whisper-typing = "whisper_typing.__main__:main"

[dependency-groups]
dev = [
  "lefthook>=1.12.4",
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "ruff>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.report]
fail_under = 100
omit = ["src/whisper_typing/tui/*", "*/__main__.py"]

[tool.coverage.run]
source = ["whisper_typing"]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchaudio = [{ index = "pytorch" }]
